Tests/LibWeb/Text/expected/HTML/HTMLObjectElement-contentWindow.txt

object.contentWindow initial value should be null: true
contentWindow.name should be the same as object.name PASS
